delay processing of queues until Nodelet::onInit() returns
This patch will delay the processing of callback queues until
Nodelet::onInit successfully returned. This will avoid race conditions
with code in onInit() and callbacks that are being called while onInit()
is being executed. It also allows to do some blocking behavior inside
of onInit() the same way we do it in a traditional node before calling
`spin()` or `while(ros::ok()) {...}`.